[A11.1.2] fixed all of the warnings.
- color.hpp initalized a in the single paramiter constructor
- Events.hpp : changed the type of  m_scroll_wheel_x and m_scroll_wheel_y to doubles (from int)
- Events.hpp : fixed the broken formmating when a non-handled event prints its ID
- font_options.hpp : changed the type of line_hight,x,y to float (from double)

+++ sdl_3_ge/header/font_options.hpp
@@ -3,7 +3,7 @@
 namespace SKC::GE {
 
 	struct font_options {
-		double line_hight, x, y;
+		float line_hight, x, y;
 		color color;
 
 		enum line_hight_mode : uint8_t {
